refactor: keep only upload statement
